tripinfoappWEB/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.0.2
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
